funcsearchindex

+++ student.cpp
@@ -1,7 +1,7 @@
 #include "student.hpp"
 #include "db.hpp"
 
-Student::Student(std::string name, std::string surname, std::string address, unsigned int index, size_t pesel, std::string gender)
+Student::Student(std::string name, std::string surname, std::string address, size_t index, size_t pesel, std::string gender)
     : name_(name), surname_(surname), address_(address), index_(index), pesel_(pesel), gender_(gender) {}
 
 void Student::setName(const std::string& name) {
@@ -13,7 +13,7 @@ void Student::setSurname(const std::string& surname) {
 void Student::setAddress(const std::string& address) {
     address_ = address;
 }
-void Student::setIndex(unsigned int index) {
+void Student::setIndex(size_t index) {
     index_ = index;
 }
 void Student::setPesel(size_t pesel) {
@@ -32,7 +32,7 @@ std::string Student::getSurname() const {
 std::string Student::getAddress() const {
     return address_;
 }
-unsigned int Student::getIndex() const {
+size_t Student::getIndex() const {
     return index_;
 }
 size_t Student::getPesel() const {
